Bump to oef sdk 0.6.1 and refactor controller agent to v1
