+++ tac/agents/v1/base/interfaces.py
@@ -102,7 +102,7 @@ class ControllerActionInterface:
         """
 
 
-class OEFSearchReactionInterface:
+class OEFReactionInterface:
     """This interface contains the methods to react to events from the OEF."""
 
     @abstractmethod
@@ -136,7 +136,7 @@ class OEFSearchReactionInterface:
         """
 
 
-class OEFSearchActionInterface:
+class OEFActionInterface:
     """This interface contains the methods to interact with the OEF."""
 
     @abstractmethod
