+++ tests/conftest.py
@@ -22,12 +22,13 @@
 import inspect
 import logging
 import os
-import subprocess
+# import subprocess
 import time
 
 import docker
 import pytest
 from docker.models.containers import Container
+from tac.helpers.oef_health_check import OEFHealthCheck
 
 logger = logging.getLogger(__name__)
 
@@ -66,10 +67,12 @@ def _wait_for_oef(max_attempts: int = 15, sleep_rate: float = 1.0):
     while not success and attempt < max_attempts:
         attempt += 1
         logger.info("Attempt {}...".format(attempt))
-        oef_healthcheck = subprocess.Popen(["python3", ROOT_DIR + "/sandbox/oef_healthcheck.py", "127.0.0.1", "10000"])
-        oef_healthcheck.wait()
-        oef_healthcheck.terminate()
-        if oef_healthcheck.returncode == 0:
+        # oef_healthcheck = subprocess.Popen(["python3", ROOT_DIR + "/sandbox/oef_healthcheck.py", "127.0.0.1", "10000"])
+        # oef_healthcheck.wait()
+        # oef_healthcheck.terminate()
+        oef_healthcheck = OEFHealthCheck("127.0.0.1", 10000)
+        result = oef_healthcheck.run()
+        if result:
             success = True
         else:
             logger.info("OEF not available yet - sleeping for {} second...".format(sleep_rate))
